Extendable pool allocator
The extendable pool allocator behaves like a pool allocator (can only allocate
elements of a certain size), but can extend itself automatically when it runs
out of memory by calling 'mbed_ualloc' to get space for another pool. The
implementation is designed to be reentrant and to minimize the time spent in
critical sections.