Changed to a slightly better API
The double step construction (allocator and then init()) in ExtendablePoolAllocator
is needed because init() allocates memory, which could fail, so it needs to return a
status, which is impossible in a constructor. This commit moves the pool construction
arguments from the constructor to init(), making the purpose of init() a bit more
clear (and saving 4 bytes of storage in the process).