Platform support for mbed-util
The initial aim for this library was to be independent from mbed-core and mbed
in general, but that has a number of issues:

- mbed-util needs a runtime error function. The mbed implementation of the runtime error
  function (https://github.com/ARMmbed/mbed-sdk-private/blob/master/source/error.c)
  depends on the target support for STDIO messages.
- mbed-util needs an assert function. The mbed implementation of the assert function
  (https://github.com/ARMmbed/mbed-sdk-private/blob/master/source/assert.c) calls
  'mbed_die', which is mbed specific.

To address this, there's now a 'mbed-util.h' header that contains the necessary
functions and an implementation that gets included only for mbed targets
(in mbed-util_mbed.c). Other implementations (for example POSIX) can be added on
top of this. Eventually I expect 'mbed-util_mbed.c' (and other similar files
that'll likely be added) to be spawned into a separate repository.

With this pattern, it's still possible to use this library in a POSIX environment
(for example) with relative ease (implement mbed-util for POSIX and add a
target POSIX dependency).

If this gets in, the corresponding functions in mbed-core should be expressed
in terms of functions in mbed-util, to avoid code duplication.