Fix CriticalSection lock for nrf51 targets.
The previous implementation of CriticalSectionLock for nrf51 was only
relying to softdevice primitives sd_nvic_critical_region_enter and
sd_nvic_critical_region_exit .
If the soft device is not enabled, these functions have no effect. This
means that CriticalSectionLock was not locking anything if the softdevice
was not enabled.

This fix allow the usage of CriticalSectionLock, even if the softdevice is
not enabled. If the softdevice is not enabled, all irq are disabled until
the destruction of the lock.