Fix uninitialized test for use with mbedgt
mbedgt expects the serial connection to be up the whole time from the
start message to the result message. The previous implementation of the
test printed messages across a reset, effectively killing the serial
connection in between.

Now the messages are printed all at once when the test succeeds. No
message is printed at all if the test fails, which will trigger a test
timeout.

Fixes: 03324ad ("Add support for uninitialized data sections")