Comment out assertions in critical.c
For the moment assertions need to be commented out as this code might
run when the libc initialization has not completed yet. In those cases,
an uninitialized stderr pointer will trigger a bus fault / uVisor access
fault as it points to flash (`__sf_fake_stderr`).

For reference:
ARMmbed/mbed-drivers#176 Printing to stderr attempts a write into flash